requirements.txt
numpy>=1.20.0
torch>=1.10.0
matplotlib>=3.3.0
pillow>=8.0.0
